tools/scraper/requirements.txt
# OpenSlots Scraper Dependencies

# Environment
python-dotenv>=1.0.0

# LLM for service filtering
anthropic>=0.40.0

# HTTP client
aiohttp>=3.9.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Browser automation (for booking systems)
playwright>=1.40.0
playwright-stealth>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Development
black>=24.0.0
ruff>=0.1.0
mypy>=1.8.0
